Add first version of `search` API (temporarily implemented with PG)
Nothing of this is in its final form, especially not the search
implementation. But I think it's useful to have separate types for
the search results instead of using the existing `Realm` types. The
search index might not store all data that our normal `Realm` is storing
and we don't want to force ourselves to always store all data in both
locations. Additionally, we might want to send additional search
related data in the new types, e.g. match locations in metadata.