Use relay default (relative) artifact directory
Now the query type definitions are generated into a `__generated__`
folder next to the file defining the query. This has the advantage of
not requiring `../../../../` in most of these imports. We also went
with `__generated__` now because it might be useful to have a clear
indication of the generated nature of the file. Finally, the relay
compiler has a bug right now that makes using custom directories pretty
annoying. So it's easier for us to switch.