Simplify `floof run clean` even more
Newer versions of `find` actually have a `-delete` command,
making it so that we don't even have to worry about the things
in the last commits.
If an environment we care about doesn't support it (yet), we can still
revert this commit, of course.

Note that `-delete` implies `-depth`, not `-prune`, so technically
we are now looking at more files, again. ¯\_(ツ)_/¯
This also necessitates switching from `-name` to `-path`
so that we also match the contents of the `__generated__` folders,
and subsequently delete them.
This is because `-delete` might not delete non-empty directories.