Change `rauta` API to make it more idiomatic for JS/TS
The previous API was created with my brain still thinking a lot like
I was programming Rust. This resulted in the `Prepared` generic type,
which in turn meant that I had to employ a trick to have existential
types and yada yada.

It's way easier if the `prepare` function returns the `render` and
`dispose` functions: that way, those arrow functions can just refer to
prepared values, without needing to "pass them through rauta". This
simplifies a lot of code.

This commit also removes one otherwise useful feature of rauta: the
URI matching algorithm. This too resulted in stuff being passed
randomly to some functions. Also not a great API. Unfortunately, in
this first commit, every route has to match against the URL very much
manually. This is not great and will be improved soon.