Improve pagination of "my videos" by using routes (query params) for it
Before, the pagination information was purely stored in React state.
This had several disadvantages, like: reloading the page (to get the
newest information) would show the first page again. And if you clicked
on a video on the second page, but then went back via browser
navigation, you would land on the first page again.

To improve the situation and to get rid of `useQueryLoader`, all of
this is now simply handled by routes. All pagination information is
stored in query URL parameters. Care has been taken to make the query
URL part fairly short and "easy to read" instead of just cramming a big
JSON into it.