Add `RootLoader` component to be used by almost all routes
This component is `Root` combined with `QueryLoader`: it loads a query
and renders `<Root>` with the given contents. The main reason for doing
this is to enforce a consistent component tree on all routes. Before,
there were many different places where `Root` would be rendered, which
leads to the header being completely rerendered on route changes. Apart
from the waste of computing, it also meant that input focus was not
kept when changing the routes -- very important for searching.

But the hope is, that this component also shortens and improves route
code.