Debounce search queries
This reduces server load and makes the page less sluggish. To be honest,
I have no idea why it is sluggish as everything goes through
`startTransition` and this is the prime example for that feature.

That said, at this point I am just trying to make the current fairly
simple solution more usable. I would like a very different solution
anyway.