Add basic favicon support
Admins can configure a single SVG file to be used as favicon. This works
well with all modern browsers, i.e. not with Safari and IE. Safari
requires a separate PNG file. We can add support for that down the line,
but having working favicons for most browsers is already an improvement.

Specifying a favicon is required. If there is no good one available,
you can always use the logo to make Tobira start. Though that's usually
not advisable as most logos are too complex.