Improve design of user info in header for logged-in users
On mobile, the icon changed to have a user with a green checkmark next
to it, better indicating that the user is logged in. On desktop, that
icon is also shown with the text "logged in as: <name>" left of it. The
text and the icon both make the user menu appear.

This is better than the previous version in my opinion. For one, I think
it looks fresher. But more importantly, the user name being in this
button box was weird, and the user menu touching that button then is
also weird. Since the user box was limited in size, but the size of the
menu was independent, it would often be a size mismatch which looked
really ugly. I don't think the original design was fixable, really.
That's why I implemented this now.