Show user information on "not found" page
This page means that routing has gone wrong, but showing it does not
mean the server is broken. So we can safely fetch user information. I
also checked: the main part of the page is shown even if the user info
is not loaded yet since it does not suspend.