Execute `onClick` handler in `Link` before our own handling
This fixes a previous bug where the onClick handler was not called for
clicks with ctrl key pressed or similar stuff. But also, this allows
to do stuff like `history.back()` to execute beforehand. This is used
by the login page.

The other TODO was removed because this would require us to know whether
"back" was indeed the my videos list. This information is not natively
accessible. And in any case, the user should just use the back function
to preserve scroll position. Finally, pagination and filters are all
stored in query parameters now, so that's not really a problem anymore.