Fix laggy behavior due to "onScroll -> replaceState" handler
For me it was fine, but for Julian it was lagging. And 200ms debounce
is totally fine. It is actually set to the correct value in `goto` and
`onbeforeunload`. Unfortunately, we cannot update it in `onPopState` as
`history.state` is already the new state. For that reason, we still
need the `onScroll` handler.