Make sure only one deployment job is running at the same time
This should fix timing issues where parallel jobs interfere. Most
commonly: if a PR is pushed to and merged right afterwards, the test
deploy is running when the "remove deployment" workflow is started.
That often results in the database being unable to be dropped because
it is used. So the "remove deployment" workflow should wait for the
test deployment workflow.

Of course, that doesn't solve everything. If the "remove deployment"
workflow is started before the "deploy" WF, then already merged PRs
are still deployed. But that's unlikely as we always wait for CI to be
green before merging.