Retrieve user data from partially errored API responses
This is mainly useful if the error is "you have to login to do this"
and the user box shows loading. With this, it now shows the login
button.

Ideally, if the user data cannot be retrieved, it should show an warning
triangle in the user box instead of the loading indicator. As in, we
want to distinguish "unknown" into "loading" and "error". But this can
be done later.