Partially revert "dynamically load player JS code" to fix bug
This was done in a47f1abbb1084f8d3aacef6d719ec4e269896185 but caused
super breaking bugs in release mode where the asset files have a hash
embedded in it. I have plans to properly fix that, but those won't
land until the Opencast summit. So to make the test deploy work during
that time, I just not load the extra files dynamically for now.