Add event ACL check to search
Only events the user can access are returned by the search API. Sadly,
Meili only supports case-insensitive filtering, so our workaround is
to just hex encode the roles. This is kinda meh, but isn't really all
that bad. But I already sent a feature request to Meili. We will see
about this in the future.