Fix `search::queue_many` by ignoring unique constraint conflicts
To be honest, this fast `copy in` way of doing this is quite involved
by now. I'm not sure if it's worth it to have this code complexity. I
mean... now it works. But if it ever causes any problems in the future,
feel free to just replace it with a simple loop and `insert` statements.

I actually measured this: upserting 500 items takes 8ms with this
`copy in` solution and 90ms with a prepared `insert into` statement and
a `for` loop. So it is a more than 10x improvement. Still, the slower
solution is unlikely to be a performance bottleneck anytime soon.