Add `auth.logout_link` configuration to make logout button a simple link
Before, our logout button just sent `DELETE /~session` and our
assumption was, that the reverse proxy could just intercept that and
do whatever is necessary. However, in some systems like Shibboleth,
users actually need to visit another domain. The reverse proxy can't do
that because the relevant cookies for that other domain are not
included in the request.

So a configurable logout link is an easy and straight forward solution.