Update to `@types/react[-dom]` 18
This came with a few breaking changes. See this for the full
explanation: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/56210

The most impactful for us is that `React.FC` does not have implicit
`children` anymore. So all components that use children, have to declare
them explicitly. This is a lot of changes, but a net win for type
safety.

Also, due to some changes, `TResult` (what `t()` returns) was not
assignable to `ReactNode`. This was fixed in `react-i18next`. Thus,
that package is also updated.

Further, `{}` was removed from `ReactNode`, which made some `Trans`
usages not compile anymore. I opened this issue:
https://github.com/i18next/react-i18next/issues/1483
This is not completely fixed yet, but by just changing one line, we can
make it compile already.