Make Plyr work with HLS streams
The behavior for >1 tracks is still unclear and there is no quality
selector in HLS mode. Regarding quality selector, we might want to
use the parsed qualities from the h3u8 file:
https://github.com/sampotts/plyr/issues/1741

However, we still want to keep the automatic quality choosing ability
of HLS. Seems to be a bit complicated and hacky.