Rename `events`/`series` API endpoints to `all_events`/`all_series`
This is a step to a more consistent API. Currently there is an
`event(id: ID)` endpoint that returns by ID. We can't have the same
for series as there plural and singular are the same. I also think
including "all" just makes it way clearer what happens. Aaaand these
APIs are probably removed in the future anyway.