Improve handling of "no read access" for the video page and video blocks
Instead of reporting this access "problem" through GraphQL errors, we
model it as part of our schema. This is actually something that relay
recommends: https://relay.dev/docs/guided-tour/rendering/error-states/#accessing-errors-in-graphql-responses

Catching normal GraphQL errors for this purpose would be quite
elaborate, so we gladly model it as part of the schema.

Unfortunately GraphQL, Juniper and Relay all make this harder than it
should be (due to bugs or... questionable design).