Introduce an additional method of authentication for external tools
The purpose is to avoid external tools having to invent users
and somehow "smuggle them past" the login proxy in order to use
APIs provided by Tobira.

The mechanism goes like this: You tell both, Tobira and your external
tool about a shared secret, which the latter sends to Tobira
in a header (which the login proxy leaves untouched) on every request
that needs an authenticated context.

We kind of went along the path of least resistance here; both,
in terms of design and implementation. The only practical purpose
it has right now is to allow Opencast, specifically, access to the
`mount`-API. Nothing else is guaranteed to work with the system,
at the moment.