Use `-bw-contrast` color instead of black/white where appropriate
There is one not so great part of this: sometimes we use the bw-contrast
of the base color, but the background color is the `-darker` version
or so. Which means we might get a slightly worse contrast. That said,
often that is the `:hover` case and it might look weird to change the
text color then.