Simplify logo margin code by requiring margin in the image
The previous system did not work well for some logos (e.g. Bern), did
not allow for separate horizontal and vertical margin values and forced
people to set a strange number in our config file. With this commit,
we just expect the appropriate margin to exist in the image itself.
This has many advantages and works with all kinds of logos, but has
one slight disadvantage: the `<Link>` surrounding the image have a
larger clickable area now. Maybe a problem, maybe not.