Use nonce with emotion-js to make it work without `unsafe-inline` CSP
We want to remove 'unsafe-inline' from the `style-src` CSP. We can't yet
do that because Paella emits inline CSS. But this already fixes all 
other sources of inline CSS by using a nonce.

Note: you might wonder how this nonce thing helps with anything since
the nonce is sent inside the HTML file. And yes, if an attacker also
injects JS code, then they can read the nonce and also insert CSS that 
way. But we still prevent the case where the attacker is only able to 
inject CSS (think a comment).