Do not include `nonce-123` in CSP header
Previous commits implement the nonce system for use with our own inline
CSS blocks and with emotion-js. However, Paella requires us to allow
'unsafe-inline' anyway. What we didn't expect is that when 'nonce-123'
is included in the CSP header, browsers then disregard the
'unsafe-inline'. Chrome message:

> Note that 'unsafe-inline' is ignored if either a hash or nonce value
> is present in the source list.

So since we need 'unsafe-inline' still, we have to remove the nonce
value. We leave all the nonce-related code though, to allow for easy
re-enabling once Paella is ready.