Fix and improve host realm queries
The host realm queries reported realms more than once if the
corresponding entity was included in a realm more than once.
(For example an event could be included directly because it is
featured, but the series is included as well.)
The easy fix would be to throw in a `distinct`, but these queries
actually perform ever so slightly better, and I happen to think
that they capture the intent better as well.