Link search events to the proper video page under their host realm
If they only have one, that is. Before this patch search events
linked to the "direct video page," and for now we continue to do so
if there is more than one host realm. (Or less than one, but that
shouldn't be possible, since that means the video is unlisted,
and should not appear in search.)