Add utilities to more easily and robustly create types from DB queries
This was a pain point for a long time. We decided to go with the rather
low level `postgres` crate. It supports basically everything we need
and is well tested. However, using it basically always resulted in
having to keep track of indices manually. E.g. the `from_row` function
of many types had a ton `row.get($n)` where `$n` was some integer.
Adding a new column to the `select` part of the query would often result
in changing lots of these indices. It just felt so low level, error
prone and unnecessary.

Of course, `Row` always allowed us to get the row value by name. But
that only helps a bit, as it's still easy to introduce typos there or
forgetting to update a string. It is also slower, but it's unclear how
much of an impact that would have been.

Then there are crates like `sqlx` which are super cool by obtaining
field and type information about the query at compile time and return
an appropriate anonymous struct type from the query. It is cool and
Tobira might still want to use that in the future, but it does not
solve all of our problems. In particular, I don't see a good way to
not having to list all columns in every single query.

So I ended up with this home grown solution that goes the middle way
and supports more use cases of Tobira. It is not type-safe as it does
not get the type information of selected fields (that would require
talking to the database, very much out of scope). But it does get rid of
manually tracking indices and provides some helper stuff to repeat as
little as possible.

Maybe it's overkill, maybe we'll find a stupid limitations of this very
soon, maybe in 5 years someone will be angrily shout "who wrote this
shit?!", or maybe it will work great.

It still has two limitations that I'd like to still address in the
future:
- `FromDb` types cannot have arbitrary SQL expressions.
- `from_row` requires ownership of the row.