Move SQL logic for loading search data into SQL views
The initial motivation for this change is not that great: my new
`FromDb` trait and friends did not play well with the queries that were
used by `search` items. I could have made the utilities more flexible
(and this might still happen), but it was easier to just move the logic
into a view.

However, I actually think this is not a bad thing to do. More SQL in
`.sql` files that is always executed, not just on specific branches.
And when manually working with the DB, it's also very helpful. And it
makes the improvements in one of the next commits possible.