Fix bad quadratic performance issue in search_events
Not sure how we did not notice this earlier, but the previous version
of the query performed really badly when there are lots of events and
blocks. It had a quadratic runtime, running one operation
|realms| * |events| many times. On our test deployment, it means
receiving all events that way took roughly 1 minute.

This new version is not perfect either, but it's a lot better (300ms
for the same data). Looking at the query plan, I think we could still
improve quite a bit of stuff, especially for the `where id = any($1)`
case, but I don't have time for that right now.