Make `FromDb` utility work with borrowed rows
This is important when two `FromDb` types are selected from one query
as we don't want to clone the `Row`. Unfortunately, we cannot properly
express this in the trait system due to the lack of GATs. So instead,
the `RowView` is just an anonymous type now, not being mentioned in
the trait at all.