Add ability for realms to receive name from one of their blocks
Only video and series blocks can be used for that purpose. And then, the
title of the video/series is used as realm name. If the video/series is
removed or "waiting", the realm has no name which is shown as "missing
name" at most places in the UI. We decided that distinguishing between
all the different reasons for a missing name is not worth it for most
cases. This information is only useful to moderators and they can still
check the realm settings to find out. Also, those cases should be very
rare and/or temporary.

But this way, we can almost keep the existing `realm.name` API, only
that the field is now nullable and is null in more cases than just the
root realm. There are still a few moderator-accessible cases in the UI
without `?? <MissingRealmName />`, thus showing an empty string. It
looks a bit awkward, but it's just for moderators and again, rare. It
can still be changed later.

There is a new API though, to get the raw information about a realm's
name. It's way more involved to get information of, and it unfortunately
also causes a separate SQL query.

Also, a `renameRealm` mutation was added as the existing `updateRealm`
had to be changed anyway and a more specific mutation seemed like a
good idea to me.

The UI for changing the realm name in the realm settings changed
completely and allows you to select a block. Additionally, the UI
disallows you to remove a block used as a name source. The DB checks
these constraints (and many more) anyway.

The "new realm" page was not changed as it still works and it can be
improved later.

Lastly, the migrations were moved around quite a bit.