Update to MeiliSearch 0.27.2
Nothing interesting about the update. Hopefully we can update to 0.28
soon -- we are only waiting for `meilisearch_sdk` to be updated. That
could take a while though, as 0.28 brought a lot of API changes.

The `status` command now doesn't show the "created" and "modified" dates
anymore as the documentation is really unclear on how to get them and
what they represent.