Move assets for development & test deploys to `util/dev-config`
Having most of these in the `backend` folder was always kind of wrong
and annoying. This cleans the backend folder.

One problem is that after the move, Tobira wouldn't pick up the config
file automatically, requiring you to pass it via arg. However, we
decided that putting the new (weird) path into the search paths in debug
mode was fine. We couldn't come up with a different solution that
provides the same level of DX.