Add `check-system` script to make sure all required tools are installed
I think this should help people new to Tobira development a lot. The
script also prints some hints for how to install for the things that
might not be entirely clear (Rust, penguin, watchexec).

One improvement would be to also check the version of all tools to make
sure that's compatible. But this is already a good start.