Remove animal name from version identifier
The idea was cute and we wanted it for "marketing", but it's also not
really useful and might even confuse (as it's just the same as the first
digit). We can still add animal emoji to the title of GitHub releases
for fun, just not in the actual version identifier.