Remove release overview table
I think this table is still useful. And we could completely
auto-generate a new row by checking pull request tags. We want to tag
PRs appropriately anyway to get good automatically generated release
notes.

However, writing this automatic generation takes some time we don't have
right now. And since not everyone is a huge fan of the table anyway,
removing it is going the path of least resistance. We can still add this
or similar things in the future, after releasing 1.0.