Remove Plyr player and always use Paella
This is done for two reasons:
- Plyr does not support "jump forward/backwards" UI buttons, but only
  works with keyboard shortcuts. And having buttons is a strong
  requirement of some people.
- For consistency reasons (to not confuse users with two vastly
  different players)

Personally, I would prefer to keep Plyr and maybe even find an
alternative to Paella for dual stream videos. I think Paella is lacking
in many regards, technically and design/UX wise. And while a new, good
design has been proposed and will be implemented, this will likely still
take a long time. And it would still leave all the technical problems
to deal with.

But I see that Opencast is very closely connected to Paella, for better
or worse. (Likely mostly for worse...)