Properly queue events for reindexing via DB triggers
This moves the all remaining "queue for index" logic into the DB. This
is useful as we don't have to remember to queue in every place in Rust
code, where some data is changed. It's all handled transparently by the
DB.

This commit also fixes a few situation where queuing an event was
forgotten before (when a name of a host realm's ancestors changed).