Remember ID mapping for Opencast items forever to keep URLs stable
It can happen for various reasons (most notably live streams) that an
Opencast event is retracted ("deleted") at some point but is later added
again. Since the URL to that event contains its ID (our ID, not the
Opencast ID!), this would lead to broken URLs. But we want URLs for one
event to work forever, even if the event is deleted in between. So we
remember any ID mapping we ever did in an extra table.

As the comment in the migration describes, I first tried to implement
this via a 'deleted' state, but this turned out to be a bad idea. Not
only resulted it in quite a big migration (because we can't `alter type`
in transactions), but more importantly, it required adjusting a lot of
other logic. Worse, it needed adjustments in places I did not predict
or thought of. So it seemed very fragile. And indeed, there are some
articles arguing against soft-deletes, especially because of the loss
of delete semantics. E.g. https://brandur.org/soft-deletion