Improve "currently playing" indicator for videos in grid view
The previous version's main problem was that the overlaid icon clashed
badly with the icon of the "no thumbnail" fallbacks, which are used
for almost all live videos. But it also clashes with many normal
thumbnails.

This new version has a play icon next to the title (slowly pulsing) and
greys out the thumbnail only by a little bit. I'm still not super happy
about this and I think it should still be improved in the future, but
this change already improves it.