Pass all tracks to Paella and pick audio source at random if necessary
This changes how we pass tracks to Paella. For one, we now pass all
tracks, even those with non `presenter` or `presentation` flavor. We
haven't seen such an event yet, but this seems like a useful change
anyway.

Also, in some cases we need to tell Paella which stream should function
as main audio source. We don't know but we just pick one at random.