Make event selector in video block editor searchable and more usable
The selector before was very ... not great. This replaces it with
something where the user can search through all events via our search.
One interesting consequence is that video blocks can now not be edited
anymore when the search is unavailable. But as that's an error condition
anyway, it's fine.

Apart from the better UX, loading the edit realm content page also
takes less time as we don't have to load all events anymore. We still
load all series, but that's already less data.