Increase limit of `searchAllEvents`
This is an ad-hoc improvement to make it less likely that users will
run into the problem of being interested in items that are not included
due to `limit`. I don't think it will really be a problem as users
should just add more specific search terms. But yeah, it's not super
optimal.

And either way, we might make the limit a parameter to the GraphQL API
at some point.