Statically link `libssl` for better binary compatibility
This is only used by the HTTP client of `meilisearch_sdk`, i.e. to
communicate with Meilisearch. In most instances, Meili runs on the same
machine without HTTPS, so this is basically never used. In particular,
Tobira cannot listen on HTTPS directly (you should use a reverse proxy
for that). And to communicate with Opencast, rustls is used. Ideally,
meilisearch_sdk would also use rustls and in the future we can do that.

But for now, we want to statically link libssl to make it easier to
deploy Tobira. The main argument against static linking is that an
administrator cannot deploy libssl security updates without updating
Tobira. Yes, but that's already true for 99% of Tobira's dependencies.
So just dynamically linking this one library isn't making a big
difference. Also, as mentioned above, libssl is not even used in most
situations. And if it's used, there is no user interaction.