Add structured video data on video pages
This should improve finding videos via Google or other search engines.
Only video pages (not the video block, for example) emit this structured
data: it's a safe start as I'm not sure how Google would display the
result if there are multiple videos on the page. And in any case, we
prefer people visiting the dedicated video page.