Improve GraphQL API handling
This commit basically implements `juniper_hyper::graphql` manually. This
is done in order to improve debug outputs, but also to make it possible
to retry executing a query if the DB transaction didn't commit
successfully. The latter is not part of this commit.

This changes a few things:
- We don't support `content-type: application/graphql` anymore. But we
  didn't use that anyway, our API is implementation detail, and there is
  not really a reason to not use `application/json`.
- We don't support batched requests anymore. We didn't use that and it's
  a non-standard extension by Apollo.
- The JSON response is not pretty printed anymore. `juniper_hyper`
  actually always pretty printed... an interesting choice.

And well, in any error case, the GraphQL is pretty printed (with debug
log level).