Add generic `PopOver` component and use it in user menu and date details
For upcoming commits, I need this in more places, so I guess one can
just do it properly and create a generic version. Which this commit
does. It works for simple tooltips (like the date details for a video)
but also for "open on click" larger menus like the user menu. Both
use cases now use the new component.