Fix DB migrations when DB is empty
In #602 a deadlock bug was fixed but it was not tested with an empty DB
where locking the table obviously fails. This commit fixes that problem
by just creating that table before anything else. It is an easy fix but
I'm pretty sure it's correct.

Consider all orderings of the two operations:

    A create if not exists
    A lock
                                B create if not exists
                                B lock

    ------------------------------------------------------

    A create if not exists
                                B create if not exists
    A lock
                                B lock

    ------------------------------------------------------

    A create if not exists
                                B create if not exists
                                B lock
    A lock

Regardless of whether the DB is empty or already has __db_migrations,
this never leads to a deadlock.