Add cancel button to video uploader
This adds an `abortController` to the uploader, which can be called by
pressing a `cancel` button when a video is being uploaded. It is also called
when a user leaves or refreshes the upload route. The controller then
aborts the current `XMLHTTPRequest` and rejects the promise it was being called from.