Add cancelled-state to uploader
This adds a "cancelled" upload-state, which is set when an upload is
manually cancelled by clicking the `cancelButton`. In this state,
the metadata selection is hidden and there is a button which rerenders
the `select files` dialog by resetting the `files` state.
This also adds a `POST /discardMediaPackage` call to clean up the state
on the Opencast side.