Add `auth.mode = "opencast"` to authenticate via Opencast users
I think having this mode is very useful for multiple reason. For one,
some organizations want exactly this auth-system and now don't need to
build it themselves anymore. We wanted to be auth-system agnostic in
Tobira, but I think it's fair to have an exception for Opencast as that
is _the obvious_ system to support out of the box. Communication with
Opencast is setup anyway.

Moreover, I think this is a great starting point for anyone setting up
Tobira. Because then, it's super easy to get Tobira auth up and running.
Setting up the auth otherwise is usually fairly involved. And until
then, you only have an empty Tobira and can't even add pages or blocks.
So I think this vastly improves the initial setup experience.

We are careful in describing this feature: there won't be any
configurable behaviors about this. We just take the most obvious route
and take exactly the data Opencast provides. Anything else needs to be
done by writing your own login handler.