Use darker grey for author line
On video pages, the selected video in the current series has a darker background
than the others. This makes the author and upload and starting time line
a little harder to read and the contrast ratio would not pass accessibility
checks.
This will be changed with the upcoming redesign, but as an intermediate
solution, this commit changes the used color for these lines to a darker
shade of grey.