Improve `db clear` by dropping all objects instead of whole schema
The old version would fail in some situations when the db user was not
allowed to delete the schema. We don't want to delete it anyway, but
just everything in it. So doing exactly that is more straight forward
anyway.

Unfortunately, it's a bit more involved as we have to list all objects
to be deleted. We do this by using `information_schema` but also
`pg_catalog`, unfortunately. The latter is not standard SQL and can
potentially change in future versions. Though those tables have changed
very little in the last few years. Our queries are likely to still
work in future versions of PostgreSQL. And this is just one helper
sub-command anyway.

The other question is: does this indeed delete all kinds of objects in
the DB? Well, probably not, but I'm fairly sure it deletes all the ones
that our migrations create. I tested this a couple of times, running
`db migrate` and `db clear` back to back and all works as expected. I
tested with PG10 and PG15. I also checked in a GUI tool and didn't see
any remaining objects of interest.