Remove disabled buttons from accessibility tree
This removes the disabled buttons in `My videos` from tab order
and screen reader exposition.
Reading around a bit, it seems that having disabled buttons
at all might not be the most accessible solution, but I think
in this case it still makes sense. Another option would be to hide
the buttons completely when disabled.