Rerender the current route on every navigation
Due to how React works, this previously only happened when you
actually visited a new route. Visiting the same route would **not**
rerender and thus also not reset any state.

This is unexpected for example when you click
the "upload video" button in the users menu while already
uploading something.