Replace own CSS-based popover with components based on `floating-ui`
I really liked the previous solution but unfortunately there were some
bad limitations to that approach. Tooltips and popovers that overflow
the viewport are very annoying and are especially problematic for
accessibility. So in the end I decided to use a library that has this
all figured out basically. `floating-ui` seems like that library.