Solve merge conflicts
There were some larger changes done to the `UserBox` in
the most recent PR, mainly the addition of the `floating-ui`
package and usage of a new `FloatingMenu` for the user menu.
As this made my own changes a little hard to incorporate without
redoing almost everything, I decided to leave out the `FloatingMenu`
in this particular file for now, but I might still add it later on.