Fix Paella not playing any videos by using 1x1 black image as preview
This fixes a critical bug where no videos could be played. Paella
started making the preview image mandatory in 1.11 or 1.12 which broke
all of our videos. This now sets a black 1x1 PNG as preview image for
all videos. We could use the thumbnail instead, but still need a
fallback for videos without thumbnail. And since using the thumbnail
is a user-facing change and the next release should just be a bugfix
release, black it is.