Fix player for live streams that end more than 25 days in the future
We want a function to trigger when the live stream is expected to be
over. For that we use `setTimeout`. As I learned today, if you pass a
number larger than 2^31 to `setTimeout`, it will execute the handler
immediately. This leads to rerendering as fast as possible, which is
bad. That number of milliseconds is roughly 24.8 days.

I doubt anyone in production encountered this as livestreams that are
scheduled to be longer than 1 days are very very uncommon (in fact,
impossible to schedule via UI).