Add series to the search index
There is a new Meili index which contains series data. It is updated
and rebuild as with the events index. The index version was bumped so
that everything us automatically rebuilt when updating to a newer 
version. Technically, it's wasteful as the realm and event indexes
don't need to be rebuild. But since rebuild is fast, we can just do
that to make sure admins don't forget to rebuild the series index
manually.