Fix `listed` status of series in search
We said that a series and all its videos are one unit and are all either
listed or all unlisted. Specifically, if one video is listed, the series
is also listed. It felt wrong to include those realms (where a video of
the series was mounted) in `host_realms`. So now `listed` is not simply
`!host_realms.is_empty()` anymore for series.