Use thumbnail as player preview image
We talked about this and everyone said that this is better than just
black (as before). However, it's not optimal as Tobira currently has
only one image per event that is used for all purposes. There should be
more images, larger ones for thumbnails, and bigger ones for the player
preview. But we can still do that later.