Adjust login button and header focus/hover styles
Previously, the login button would use an `ActionIcon` component for the
mobile version. But when that is placed inside a `Link` component or vice versa,
both elements would be focusable. Now the icon is simply placed inside a `div`
which uses the same styling as the `ActionIcon` components.
This also changes most focus/hover styles from boxShadow to outline for simplicity
and now uses a more subtle hover style for all interactive header elements.