Adjust language menu
The language menu icon now uses a smaller fixed size for
all screen sizes. In order to keep a uniform alignment of
all buttons inside the button container, I added a media query
to adjust the padding of the language icon on medium screen sizes
and above, as the other icons still get shrunk a few px for small
screens.
The language menu now also closes on language selection.