Add user menu entry linking to own user realm
It is shown whenever the user may create a user realm, regardless of
whether it exists yet. For one, otherwise every page load would have one
additional query checking if the user realm exists, putting extra stress
on the DB, likely mostly for nothing. Further, always showing it
improves the visibility of the feature and makes users aware of it.