Improve API error handling to properly allow inspecting DB errors
We would like to inspect database errors to match specific constraint
errors to specific user facing errors, for example. (Otherwise we would
need to duplicate that logic in Rust.)

However, before this commit that was not necessary as our `Tx` helper
would log any DB error immediately and later, the `handle_api` would
check if any DB error occurred and reply with 500 then. That's not
very useful.

With this commit, we will roll back the DB transaction if any
`FieldError`s happen during query execution. To find that out, we had
to use a lower level Juniper API and duplicate a bit of Juniper code.
However, we don't always reply 500 in that case, but reply with those
exact field errors, but no data.